[temp] Make spdlog target consistent with documentation
This is required when building Kodi with ROS 2 on Ubuntu 18.04.

Ubuntu 18.04 ships with an older spdlog, so an internal build is required.
However, ROS 2 on Ubuntu 18.04 also uses an internal build for the same
reason, causing a clash of target names.

By giving Kodi's internal spdlog a unique target name, the build succeeds.

Error was:

  CMake Error at spdlogConfigTargets.cmake:42 (message):
    Some (but not all) targets in this export set were already defined.

    Targets Defined: spdlog::spdlog

    Targets not yet defined: spdlog::spdlog_header_only